chore(backend): clear 39 pre-existing ruff errors so `ruff check .` is green (#143)
Backend CI runs `ruff check .` but only on PRs that touch backend/**;
recent purely-frontend PRs (#141, #142) skipped the gate, so debt
accumulated and the local check now reports 39 errors against
develop HEAD.

This sweep is mechanical and behaviour-preserving:

- 14 auto-fixed by `ruff --fix` (import sorting, unnecessary quoted
  annotations, unused-import cleanup, typing.Callable -> collections.abc).
- Long lines in `scripts/seed-dev-data.py` reformatted to multi-line
  dict literals; same data, just one field per line.
- Ambiguous CJK punctuation (fullwidth `，`, en-dash `–`, multiplication
  `×`) in zh-TW user-facing strings: kept the typographically-correct
  glyph and added a targeted `noqa: RUF001/RUF002`. ASCII docstring
  uses of `–` / `×` were replaced with `-` / `*`.
- `app/worker.py`: replaced the `while True: await asyncio.sleep(3600)`
  park-loop with `asyncio.Event().wait()` — same blocking semantics,
  no spin-sleep, satisfies ASYNC110.
- `pubsub.py` shutdown try/except/pass: now `except ... as exc:
  log.debug(...)` so the failure is observable in CloudWatch instead
  of silently swallowed (S110).
- `Body(default_factory=dict)` default in FastAPI handler: kept (it's
  the canonical FastAPI DI shape) with an inline `noqa: B008` and a
  reason.
- `random.Random("…")` in seed script: `noqa: S311` — dev seed, not
  crypto; the docstring already explained the deterministic seed.
- `int(round(...))` in `simple_overlap.py`: dropped the redundant
  cast (RUF046; `round()` returns `int` when ndigits is omitted).
- Fixed a latent bug `id_factory: callable[[], str]` (lowercase
  builtin `callable`, would crash at import once the quoted forward
  ref was removed by RUF010) -> `Callable[[], str]` from
  `collections.abc`.

After: `ruff check .` clean. `pytest -q` 566 / 566 unit tests pass
(5 deselected — `test_check_migrations.py` only on Windows due to
cp950 codec on em-dash bytes; passes on Linux CI).

+++ backend/app/domain/services/strategies/simple_overlap.py
@@ -42,10 +42,13 @@ class SimpleOverlapStrategy(ICompatibilityStrategy):
     ) -> CompatibilityScore:
         overlap = _overlap_ratio(requester_focus_starts, candidate_focus_starts)
         affinity = _affinity(requester.role_label, candidate.role_label)
-        base = int(round(overlap * 80)) + affinity + 5
+        base = round(overlap * 80) + affinity + 5
         score = max(40, min(99, base))
+        # Fullwidth commas are intentional — this is the zh-TW user-facing
+        # reason string; ASCII commas read as half-width and break the
+        # visual cadence next to CJK glyphs.
         reason = (
-            f"你們的專注時段重疊度 {int(overlap * 100)}%，"
-            f"加上角色組合的隱性默契，今晚很可能合拍。"
+            f"你們的專注時段重疊度 {int(overlap * 100)}%，"  # noqa: RUF001
+            f"加上角色組合的隱性默契，今晚很可能合拍。"  # noqa: RUF001
         )
         return CompatibilityScore(score=score, reason=reason)
